refactor(core): modularize local collection registry
